@@ -15,30 +15,29 @@
   "dependencies": {
     "@mdi/font": "^7.4.47",
     "@vueform/slider": "^2.1.10",
-    "apexcharts": "^3.46.0",
-    "axios": "^1.6.7",
-    "chart.js": "^4.4.2",
+    "axios": "^1.7.2",
+    "chart.js": "^4.4.3",
     "chartjs-adapter-moment": "^1.0.1",
     "pinia": "^2.1.7",
     "pinia-plugin-persistedstate": "^3.2.1",
-    "vue": "^3.3.4",
-    "vue-chartjs": "^5.3.0",
-    "vue-router": "^4.3.0",
-    "vue3-apexcharts": "^1.5.2",
-    "vuetify": "^3.5.7"
+    "vue": "^3.4.27",
+    "vue-chartjs": "^5.3.1",
+    "vue-router": "^4.3.2",
+    "vuetify": "^3.6.8"
   },
   "devDependencies": {
     "@types/chart.js": "^2.9.41",
     "@vitejs/plugin-vue": "^4.6.2",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.3",
-    "jsdom": "^21.1.2",
-    "prettier": "^2.8.8",
-    "sass": "^1.71.1",
-    "typescript": "^4.9.5",
-    "vite": "^4.5.2",
-    "vite-plugin-vuetify": "^1.0.2",
-    "vitest": "^0.31.4",
-    "vue-tsc": "^1.8.27"
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vueform/slider": "^2.1.10",
+    "jsdom": "^24.1.0",
+    "prettier": "^3.3.0",
+    "sass": "^1.77.4",
+    "typescript": "^5.3.3",
+    "vite": "^5.2.12",
+    "vite-plugin-vuetify": "^2.0.3",
+    "vitest": "^1.6.0",
+    "vue-tsc": "^2.0.19"
   }
 }
